+++ store/src/test/java/com/zextras/mailbox/util/SoapClient.java
@@ -7,8 +7,9 @@ package com.zextras.mailbox.util;
 import com.zextras.mailbox.soap.SoapUtils;
 import com.zimbra.common.service.ServiceException;
 import com.zimbra.common.soap.Element;
+import com.zimbra.common.soap.HeaderConstants;
 import com.zimbra.common.soap.SoapProtocol;
-import com.zimbra.common.soap.XmlParseException;
+import com.zimbra.common.soap.SoapUtil;
 import com.zimbra.common.util.ZimbraCookie;
 import com.zimbra.cs.account.Account;
 import com.zimbra.cs.account.AuthTokenException;
@@ -123,6 +124,7 @@ public class SoapClient {
 		}
 
 		public SoapResponse execute() throws Exception {
+			// Cookie: ZM_AUTH_TOKEN
 			final BasicCookieStore cookieStore = createCookieStore();
 			if (!Objects.isNull(caller)) {
 				cookieStore.addCookie(createAuthCookie());
@@ -130,7 +132,7 @@ public class SoapClient {
 
 			final HttpPost httpPost = new HttpPost();
 			httpPost.setURI(URI.create(this.url));
-			httpPost.setEntity(createEnvelop());
+			httpPost.setEntity(createEnvelope());
 			try (var client = getClient(cookieStore)) {
 				final CloseableHttpResponse response = client.execute(httpPost);
 				return new SoapResponse(response.getStatusLine().getStatusCode(), SoapUtils.getResponse(response));
@@ -142,24 +144,28 @@ public class SoapClient {
 			return this.execute();
 		}
 
-		private StringEntity createEnvelop() throws XmlParseException, UnsupportedEncodingException {
+		private StringEntity createEnvelope()
+				throws ServiceException, UnsupportedEncodingException {
 			Element envelope;
-			if (Objects.isNull(requestedAccount)) {
+			Element header = SoapUtil.createEmptyContext(SoapProtocol.Soap12);
+			if (!Objects.isNull(caller)) {
+				var authToken = AuthProvider.getAuthToken(caller, isAdminAccount()).toZAuthToken();
+				header = SoapUtil.toCtxt(SoapProtocol.Soap12, authToken);
+			}
+			if (!Objects.isNull(requestedAccount)) {
+				SoapUtil.addTargetAccountToCtxt(header, requestedAccount.getId(), requestedAccount.getName());
+			}
+			if (Objects.isNull(header)) {
 				envelope = SoapProtocol.Soap12.soapEnvelope(soapBody);
 			} else {
-				final Element headerXml =
-						Element.parseXML(
-								String.format(
-										"<context xmlns=\"urn:zimbra\"><account by=\"id\">%s</account></context>",
-										requestedAccount.getId()));
-				envelope = SoapProtocol.Soap12.soapEnvelope(soapBody, headerXml);
+				envelope = SoapProtocol.Soap12.soapEnvelope(soapBody, header);
 			}
 			return new StringEntity(envelope.toString());
 		}
 
 		private BasicClientCookie createAuthCookie() throws AuthTokenException, ServiceException {
 			final var authToken = AuthProvider.getAuthToken(caller, isAdminAccount());
-			final var name = ZimbraCookie.authTokenCookieName(false);
+			final var name = ZimbraCookie.authTokenCookieName(isAdminAccount());
 			final var cookie = new BasicClientCookie(name, authToken.getEncoded());
 			cookie.setDomain(caller.getServerName());
 			cookie.setPath("/");